requirements.txt
streamlit
pillow
opencv-python
pytesseract
googletrans==4.0.0-rc1

easyocr
numpy
